Change project-level build.gradle to KTS
Change the project-level build.gradle to a KTS script so that it is easier to
work with. A "Kotlin" object is added to AppDependencies.kt so that the Kotlin
version can be declared where all versions and dependencies are declared. This
meant that the `plugins` block in the project-level build.gradle is now, as of
this commit, a `dependencies` block. This is required because `plugins` are
evaluated first and cannot access declared variables in the buildSrc but the
`dependencies` block can.