Add Jetpack Compose navigation to application
Prior to this commit, the codebase had the UI more or less set up and ready for
both the driver list screen and the route list screen. This commit adds
navigation so that clicking a driver in the driver list screen directs the user
to the route list screen for that driver.